fastapi-server/requirements.txt
fastapi
uvicorn
grpcio-tools
